requirements.txt
streamlit
pandas
numpy
spacy
pdfkit
python-docx
fuzzywuzzy
scikit-learn
matplotlib
plotly
